move most backup files back into the lib.
