+++ src/sync/statics.rs
